[expo-video][android] Fix video pausing when entering PiP and Fullscreen (#27697)
# Why

The activity calls the background event when entering Fullscreen and
PiP, which caused the video to pause. This fixes this issue. Video will
still pause when auto-entering PiP, as we have no way of knowing for
sure if the activity will enter PiP or just background. Once
`onUserLeaveHint` is available to us we will be able to fix this easily.

# How

Added a check that checks if the video is in fullscreen or will enter
PiP when deciding if we should pause the video

# Test Plan

Tested on Android 14 emulator in Bare Expo