[router]: ensure initial location can include a (group) (#27935)
# Why

Hoisted index routes inside a group will be exported as a 404 route.

FYI: This is the fix for @byCedric's 404 issue, there wasn't actually a
problem with the exporting logic.

# How

The issue is how we sort the RouteConfigs within `getStateFromPath`. A
hoisted index route will have a screen name like `(a)/(b)/index`. This
will be progressed as

```
aParts = [] // This is stripped of all groups
a.screen = '(a)/(b)/index'
```

When compared to the `+not-found` route (`bParts = ['+not-found`]`), the
`+not-found` route will have a higher priority as its more specific.
While we have logic for checking for `+not-found`, that logic occurs
after the parts length check (longest length wins).

Non-hoisted index routes are not affected by this due to the `a.screen
=== 'index'` guard. The fix is to simply extend this guard to also check
for hoisted index routes. `if (a.screen === 'index' ||
a.screen.match(/\/index$/)) {`
 

# Test Plan

- Created new unit tests
- Modified `server` E2E test to include a hoisted index route inside a
group

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>