Fix ESLint recommendations to match React team's recommendations
It is very bad that this is disabled in something that's recommended on the website. This is how React gets reputation for having "stale closures". I guarantee that there are many many many hours of lost work attributed to this choice.

If https://github.com/expo/expo/pull/26964 has no path to merging, please let's at least merge this.