[eslint-config-expo] eslint config for expo projects (#27915)
# Why

We want a generic eslint config that we could recommend for all expo
react native projects.

# How

This starts by copying the `native` config from
`eslint-config-universe`. The difference is that we won't include
prettier (the cli will have an option to install it separately), and we
_do_ include `eslint-plugin-expo`.

## TODO

- [x] remove stylistically opinionated rules

# Test Plan

Testing locally by installing it on an existing project with
[yalc](https://github.com/wclr/yalc?tab=readme-ov-file)

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).