Prevent non-async chunks to be async imported
In `metro-config/src/serializer/serializeChunks`, during chunk splitting
in our custom serializer, we allow async paths to be resolved to
non-async chunks.

This can happen if the serializer decides that an async chunk should be
in the entrypoint chunk, i.e. if it squashes away the async chunk.
Likely related to comment reading: "Prevent circular dependencies from
creating infinite loops."

The above has the unintended side-effect that an async import may then
depend on the entrypoint chunk, which causes the async importer to
reload and reevaluate said chunk.