[expo] support better hermes engine resolutions (#27994)
# Why

fixes #27941 
close ENG-11841

# How

the DefaultReactNativeHost will look up hermes if hermesEnabled=true.
since we don't call `getJSEngineResolutionAlgorithm()` to
DefaultReactHost, we will have the default behavior to try jsc first then hermes.

this pr added the `getJSEngineResolutionAlgorithm()` support and
delegate to DefaultReactNativeHost. then it would reduce some startup
time and strange adb log for libjscexecutor.so not found when using hermes.
this is sdk-50 commit only since on main i've added this already for the
work of bridgeless support.