[docs] Fix ImagePicker Sample Code in Expo SDK Documentation (#28021)
# Why

This PR addresses an issue in the official SDK documentation for
ImagePicker, where the provided sample code contained errors that
prevented it from running successfully. By correcting these errors, this
PR aims to improve the developer experience and ensure that the
documentation serves as a reliable resource for implementing the
ImagePicker functionality.

- here is the error sdk doc link:
[https://docs.expo.dev/versions/latest/sdk/imagepicker/#usage](https://docs.expo.dev/versions/latest/sdk/imagepicker/#usage)

# How

I reviewed the sample code provided in the ImagePicker documentation and
identified the errors that were causing the code to fail. After making
the necessary corrections, I tested the modified sample to ensure that
it runs as expected. The changes were made with the goal of maintaining
the original intent of the sample while ensuring its functionality.


# Preview


https://github.com/expo/expo/assets/10773980/d04661a2-3092-4324-8a33-f2559688bd2f



# Test Plan

To verify the changes, follow these steps:

1. Implement the updated sample code in a new or existing Expo project.
2. Run the project and navigate to the sectio


n where the ImagePicker is used.
3. Trigger the ImagePicker functionality to ensure that it operates
without any errors.
4. The test was conducted in a controlled environment to ensure the
reliability of the results.

OR

- Use preview and visit:
[https://snack.expo.dev/@erickuang/image-picker](https://snack.expo.dev/@erickuang/image-picker)


# Checklist


- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).