chore: relax forbidden react server API errors (#27878)
# Why

In RSC, components can be shared between server and client, the client
components may contain react hooks which will run on the client but
shouldn't run on the server. This PR reduces the errors to only throw
when APIs are used and not when they're simply imported. The import
errors still run when importing `Component` or `PureComponent`.

To fully support shared components, we'll need a follow up PR to
"optimize" server components. This should collapse unused hooks so that
they don't run, and therefore don't trigger the static error pass. This
optimization pass is a bit tricky since we have old babel passes that
obfuscate the expected syntax, ref
https://github.com/facebook/react-native/pull/43662

# Test Plan

- Updated the tests with the expected syntax.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>