[core][Android] Refactor how we stores information about exported functions (#28163)
# Why

Refactors the way we store information about exported functions.
Hopefully, fixed segfaults in `expo::MethodMetadata::convertJSIArgsToJNI`.

# How

I used to think that Java would have a longer lifespan than C++. However, it seems that C++ has outlasted Java. Even after cleaning everything, some functions are still being called on the old runtime, leading to segmentation faults. To address this issue, I had to change how and where we store information about exported functions. Previously, the data was stored inside the `JavaScriptModuleObject`, and to access it, we had to use that object. This was not ideal because we didn't have full control over the lifetime of the object. Now, the data is also stored in the module object, but each exported function has its own weak pointer. During execution, the function can handle the case where metadata was deallocated.

# Test Plan

- bare-expo ✅ 
- bare-expo with a memory profiler to check if we don't ✅