[core][updates] fix error recovery on android (#28177)
# Why

the error recovery of updates was broken since
https://github.com/expo/expo/pull/27629/files#diff-b034a3a37c87b036bb0d9c0549d6701774c79fedb4892dd44807d8947d75d4fbL27-R37.
apparently when ReactContext is available, it's too late for error
recovery because bundle is loaded and may throw exceptions already.

# How

- i feel #27629's change is correct for `onDidCreateReactInstance`. for
error recovery we sound like need to call point between
onWillCreateReactInstance and DevSupportManager available (error
recovery requires DevSupportManager to detect exceptions). this pr
introduces a new `onDidCreateDevSupportManager` for this purpose.
- also fix incorrect onWillCreateReactInstance timing on bridgeless mode
- seen a detox failed from proguard that i have to add an additional
proguard rule to fix that.