[core] fix installModules random crash (#28200)
# Why

the `TypeError: Cannot read property 'NativeModule' of undefined` still
happens occasionally. the problem is that installModules may happen
right after the block.
the `JavaScriptContextHolder` may be ready after that on the second block.
we are not able to install expo global object when
JavaScriptContextHolder is not ready

# How

it's not clear to me whether it should be an issue from react native
core. but to unblock our testing, let's add a busy waiting workaround first.