[go] enable bytecode vtab for sqlite (#28281)
# Why

enable the sqlite bytecode vtab feature for livestore integration

# How

originally i was thinking to add the feature as a builtin to all
expo-sqlite libs but the sqlite3.podspec is not published yet. let's add this for expo-go first. for people using
development builds, they could use a config-plugin to enable the feature.