[cli] bump swc to fix js debugger launching error (#28315)
# Why

fixing an error `TypeError: osascript(...) is not a function` when
pressing "j" to launch js debugger
this is a regression from #27061 and actually a bug from swc

# How

upgrade swc fixed the problem, maybe a relevant fix is
https://github.com/swc-project/swc/commit/0a7ca2f4bbfaa2573b4da6b25127064cbcfd755f?

it's near release stage and i don't want to upgrade swc 1.4. let's stick on 1.2.