[video][web] Fix `AudioContext` based audio management issues (#28224)
# Why

There were errors thrown on fast-refresh and refresh when using
`expo-video` on web. This was due to incorrect audio nodes management.
`AudioContext.createMediaElementSource` was called multiple times per
`<video>` element, which is not allowed. Also during a refresh due to
changes in `VideoView.web.tsx` the refresh would fail, because new
AudioContext was created, but the video was still bound to the previous
one.

# How

Moved the AudioContext creation into the VideoView component. This means
that each VideoView has it's own AudioContext that will stay the same
throughout the lifecycle of the component.

In external apps `expo-video` on web was throwing an error due to
checking `window && ...`. Changed this check to `typeof window !==
'undefinded`

# Test Plan

Texted in NCL on web and in a separate app.