[docs] Link to new files from native project diffs (#27754)
# Why
A user of the Native Project Upgrade Helper suggested that it would be
great to link to the new file in the diff, because sometimes (often?)
you want to just copy that whole file over. I can definitely relate to
this from past native project upgrades.

# How
This is probably not the final form.

I tried adding a simple link icon/button to the `SnippetHeader` that
would send you to the new file.

However, using a normal `A` or `LinkBase` ends up with this error:

![image](https://github.com/expo/expo/assets/8053974/15f1aa0c-4e6d-4c03-9927-1b62841daa46)

This is due to the embedded `A` tags on account of the
`PermalinkedSnippetHeader`. I can hack around it with `window.open()`,
which seems wrong. It also means the link doesn't show correctly when
you hover over it.

I could fix this by moving the permalinking to a lower level, but wanted
to check first if this is the preferred UI for such a thing.

It's not entirely different from the React Native Upgrade Advisor, which
puts the new file behind the "Raw" button:
<img width="670" alt="image"
src="https://github.com/expo/expo/assets/8053974/80d41159-7063-479f-9478-f8e91241aeb3">

# Test Plan
- [x] make sure links work on native project diffs
- [x] make sure links don't show up on other difffs

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Bartosz Kaszubowski <gosimek@gmail.com>