[video][ios] Improve audio session management (#28205)
# Why

- Fixes the audio being broken when `allowPictureInPicture` option is
initially set to `false`
- When all of the players are muted, or no players are playing audio
from other apps won't be interrupted.

# How

Used correct audio session category to fix the issues. Every time an
option, which relies on correct audio session is toggled, the
`VideoManager. setAppropriateAudioSessionOrWarn` is called to make sure
that functionality works.

# Test Plan

Tested in BareExpo on a physical iOS 17 device