[video][android] Add support for setting now playing title and visibility (#28390)
# Why

Adds support for setting now playing title and visibility for android

# How

Added a property to the video source that contains the title and second line of text. 
VideoManager was holds a weak reference in `mediaItemsToVideoSources` but no other instance was holding a strong reference. Now the player will hold a strong reference to the current video source. This should be enough as there is no VideoSource queueing added yet.

# Test Plan

Tested in BareExpo on a Pixel 8 running Android 14