[docs] Add a progress tracker for EAS Tutorial (#27971)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

Fixes ENG-11468

Add a Progress Tracker component to track progress when a chapter is
completed in the EAS Tutorial. It:
- shows a success message when the user reaches the end of the chapter
with the summary of that chapter
- allows the user to mark a chapter read/unread
- syncs the state of chapter (read/unread) with the sidebar
- adds and updates the sidebar circular progress on reaching tutorial
complition
- allows user to reset the progress on reaching tutorial complition

Also, handles Google Analytics 4 event to send a custom event when all
chapters in a tutorial are completed. The event added is called
`eas_tutorial`.

# Test Plan

To test it, see the **Summary** section of a chapter. For example:
http://docs.expo.dev-pr-27971.s3-website-us-east-1.amazonaws.com/tutorial/eas/configure-development-build/#summary.

## Preview

### Saving Progress Tracker for a chapter



https://github.com/expo/expo/assets/10234615/ced38210-e4ea-4c6e-b7d9-578f1a4ee5db



### Completed Progress Tracker and reset tutorial functionality



https://github.com/expo/expo/assets/10234615/3d671c8b-6f79-4c5c-a166-b88a03f6afbd






<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->


# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).