[dev-menu][iOS] Fix JS Debugger not detecting correct engine in new arch (#28606)
# Why

When using the new arch with bridgeless mode the DevMenu misidentifies
the JS engine as JSC

# How

When using new arch with bridgeless mode use the value of `USE_HERMES`
to determine the JS engine. As far as I know, V8 is not supported on iOS
with bridgeless so we don't need to check for that case.

# Test Plan

Run FabricTester with Hermes and JSC



https://github.com/expo/expo/assets/11707729/808fa060-4b73-489e-8d2f-5307a639175a



# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).