[av][android] Fix events not working (#28716)
# Why

This PR fixes a regression introduced in SDK 51, caused by the fact that
events sent using the wrong event emitter are not accepted.

Fixes https://github.com/expo/expo/issues/28698

# How

Used a solution similar to `ExpoBattery`, where the `sendEvent` function
from the module is passed to the calling class instance. In this case I
used additional wrapper interface, because `AVModule` is written in
kotlin and `AVManager` is in java.

# Test Plan

Tested in BareExpo in Android 14 emulator.