fix(cli): exit `findUp` methods if parent dir equals current dir (#28801)
# Why

Fixes #28799

# How

Determining if the current `cwd` is the root of a drive was hardcoded,
and not correct for Windows. On windows, the root is something like
`C:\` or `D:\`, which also cannot be hardcoded.

Instead, it now relies on `path.dirname(cwd)`, if that equals the
current `cwd` it stops. It basically means that we can't go one up
anymore.

# Test Plan

- See #28799

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).