[docs] Update templates for starting the tutorial and an Expo Router app (#28766)
# Why
The default template now has Expo Router, but the tutorial assumes just
an App.js file.

# How
Updated the command to use the blank template. Long-term, it's surely
better to update the entire tutorial to use the default template, but we
really gotta go through and test that, and maybe adjust some things to
account for the default template having multiple tabs we don't need,
etc.

Also added a callout explaining why we're not using the default
template. Maybe it's too much info, but I feel like a lot of people with
big ideas trying to learn RN are going to think "what if I want a second
screen?" so why not answer their question right away.

That lead me to also update the Router starter, as that refers to the
SDK 50 tabs template.

# Test Plan

Look at it

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>