[tools] Assigning SDK tags also on the main branch (#28894)
# Why

Currently `et publish` assigns `sdk-*` tags on npm only when running on
the release branch. There is a period of time when `main` and the
release branch are not diverged in which case we're still publishing
from `main`. It would be great to tag these versions too.

# How

- Get the proper tag based on the `expo` package version instead of the
branch name
- Always assign a tag if we're on the release branch
- Prompt what to do when we're on any other branch

# Test Plan

`et publish expo-font --assign-sdk-tag` assigned the tag properly, but
it only runs this specific task and not the entire flow, so I'll also
need to test this change once it lands on main.

Co-authored-by: Tomasz Sapeta <tsapeta@users.noreply.github.com>