Fix error: no viable constructor or deduction guide for deduction of template arguments of 'weak_ptr' (#29075)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub issues, forums posts, or feature requests.
-->
This is an error that occurs on Expo 51

# How

<!--
How did you build this feature or fix this bug and why?
-->
This is what I patched in my project to make apk building works

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->
I test with the original code and fails with `error: no viable constructor or deduction guide for deduction of template arguments of 'weak_ptr'`. And the error is gone when I replace the line with the fix.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).