chore(cli): skip starting dev server in exports (#28552)
With static rendering we changed the CLI to start a dev server so we
could ping it for resources. This prevents us from running tests in
parallel and dilutes error handling as we have to parse them back from
server requests.

This PR forks part of the metro dev server `.bundle` request into Expo
CLI so we can call it directly for SSR instead of going through the dev
server. We then skip starting the dev server in order to prevent using
any ports.

We have to leave some of the legacy SSR code in place to support the
error boundary (pending rework).

In the future we can consolidate more by removing all metro export
interface code in favor of this code path, but we don't need to do that
in this PR.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

- Consolidate export code from SPA-styled exports and SSR imports.
- Remove all SSR error translation code since we now throw errors
directly in-process.
- Remove all port freeing code since export no longer starts a server.e

<!--
How did you build this feature or fix this bug and why?
-->

Test coverage for this is very high, hopefully tests run a bit faster
now.

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>