[core] remove view wrapper from ExpoFabricView (#28829)
a trial to remove the view wrapper from ExpoFabricView

1. let native views inherit ExpoFabricView, so that we don't need
ExpoFabricView as a wrapper. this is actually done by `typealias ExpoView = ExpoFabricView`.
2. fabric will call `[viewClass new]`
since we need AppContext and ModuleHolder in the initializers, the idea
is to swizzle the `new` initializer dynamically.
3. remove `contentView` related logic
4. move `__injectedAppContext` from ExpoFabricViewObjC to
ExpoFabricView. that would make the logic clearer because we put the
code together.
5. since we swizzle `ViewManagerAdapter_ExpoImage.new()` to
`ImageView.init(appContext:)` and we need moduleHolder/viewDefintion in
the contructor (for `installEventDispatchers()`). it doesn't give us
chance to inject iMethod or iVar to the `ImageView` class for the
moduleName. this pr tries to add dedicated `ExpoFabricView.create()` and
passing viewDefinition into the class. not a perfect case but works
though.

---------

Co-authored-by: Tomasz Sapeta <tomasz.sapeta@swmansion.com>
(cherry picked from commit 8421bfb116dd3f21c8c4f66ed0fda010cdf2f078)