fix(cli, metro-config): Fix source map generation in development. (#29463)
# Why

- The source maps are off-by-1 in development and this causes debugging
and errors to be janky.
- The cause is related to Metro serializing the graph differently in
`.map` requests than it does in `.bundle` requests.
- Required change in Metro https://github.com/facebook/metro/pull/1284

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- Mutate the premodules to try and mitigate issues like this from
happening in the future.
- Detect when a serialization request is meant for source maps in
development, and return the maps from the serializer. This depends on an
upstream fix in Metro (if it doesn't land, we can try something else).
- [Update]: I've also added a new system (consider existing
implementation legacy) which injects a polyfill that is one line long.
We then update this polyfill in the serializer accordingly to ensure
source maps stay correct, even in cases such as error handling where
Metro serializes them a third way.


<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

- Added an E2E dev server test which pings the source maps and detects
if the `__env__` module is included.
- Physically attached a debugger and ensured that the sources were
aligned in development.

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>