Fix cancel button on expo-contacts presentFormAsync for unknown contacts
Currently the cancelButtonTitle isn't passed down to the controller which makes the contact form popup difficult to close as the cancel button is not visible.
This is not an issue for the new contact form as the default Cancel and Done button will be visible anyway.

This addresses the problem by calling the setCloseButton as expected