[core][ios] Fix random conversion errors when converting JavaScript floating-point numbers to `Swift.Float` (#29053)
# Why

Fixes https://github.com/expo/expo/issues/29036

Fixes a conversion error, which would be thrown when setting some
specific numbers (eq. 0.7 which is actually something like 0.6999999993)
as `playbackRate` or `volume`. Conversion fails because JS sends the
value as a Double, and for some precisions the conversion using
`floatValue as? Double` returns null

# How

Made all of the Property Float types Double.

# Test Plan

Tested in BareExpo on iOS 17.5