[docs]: Add docs for +native-intend and other linking methods (#29592)
# Why

Documents the `+native-intent` feature and documents how users can
handle deeplinks and URL rewrites within their application. This is
loosely based upon
https://reactnavigation.org/docs/5.x/deep-linking/#test-deep-linking-on-ios

The aim for this guide is to show people that `+native-intent` exists
and the problem it solves, but to avoid people being reliant on it and
over using it.