Fix @expo/vector-icons version in bundledNativeModules (#29671)
# Why

`@expo/vector-icons@14.0.1` shipped the prop-types fix. If you upgrade
from 50->51 and your lockfile already had `14.0.0`, `@expo/vector-icons`
will not get updated to the latest and you'll get:

```
Unable to resolve "prop-types" from "node_modules/@expo/vector-icons/build/vendor/react-native-vector-icons/lib/icon-button.js"
```

No idea how common this is, but it's happened to me twice now, and I'm a
little confused each time :-D.

# How

Updated to `^14.0.2` as that's the latest and there's no reason not to
have it on SDK 51.