[video][ios][android] Add support for request headers to in the video source (#29539)
# Why

`expo-av` supported adding custom request headers, `expo-video` should
do that too.

Adding a custom header to the HTML5 player is unsupported, I couldn't
find any decent enough way of doing it. A Service Worker seems like the
only solution that could work, but it's still hacky.

https://github.com/expo/expo/issues/29436
ENG-12455

# How

Implementation is mostly based on the `expo-av` implementation, on iOS
used the private `AVURLAssetHTTPHeaderFieldsKey` to set the header.
Android is a bit more complex and requires creating a data source
factory.

# Test Plan

Tested in BareExpo in Android emulator and iOS simulator. Headers seem
to be sent correctly, although we don't have any video sources that
require a custom header, so I couldn't test it in practice.