[video][ios] Add base64 certificate support (#28990)
Some of the users requested the feature of passing the certificate as
base64 string instead of a server url.

Added an additional property in the `DRMOptions` to accept the
`base64CertificateData`, when provided it will override the
`certificateUrl.

Tested in BareExpo on a physical iOS device.

(cherry picked from commit cbf10d76548dfad9aed39e31aa3425854a4c3337)