[docs] Improve GitHub pages integration guide (#29881)
# Why
I tried this last night - it worked, but some of the steps didn't quite
line up. This is such an easy way to deploy a static site that everyone
has free access to, but maybe just needs a little extra context (GitHub
doesn't do a great job of explaining it IMO)

# How
- Originally step 1 told you to "skip this step" if your repo is already
setup on Github, but really you needed to skip the first two steps. It
felt a little more confusing than it needed to be for someone who uses
Github a lot, but maybe mostly through a GUI, and maybe is already used
to regularly putting stuff on Github, so compressed that into "1 step
(if you need it)"
- Most of the other changes are at the end. The really critical bit is
that `gh-pages` is putting your exported app on a `gh-pages` branch, and
Github reads from that branch to serve it. Let's spell out the name of
the branch (it's a very specific branch name, no other will do!) and
also show the configuration that's needed in GitHub. AFAIK, that config
doesn't happen automatically.
- Finally, contextualize how it's actually a Github action that makes
your site come alive. It's the evidence that everything else actually
worked, and its easy to miss if you're not used to Github actions.

> NOTE: this doc wrote all the terminal commands in `yarn`. I wasn't
sure if we want to break out into the different package managers
everywhere, or only when that context was really important. I can update
that throughout the guide as needed.

# Test Plan

Tried this out last night, worked great.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>