[video][web]Fix `AudioContext` being created before user interaction (#29695)
# Why

On web expo-video creates an `AudioContext` after user enters the page.
On most browsers it is an issue, as AudioContext can be created only
after user interaction, this caused serious playback issues.

# How

Delayed `AudioContext` creation to start as the video is playing, or if
it is playing before user interaction (possible if video is muted) to
when the volume is changed.

# Test Plan

Tested in NCL in Chrome, Brave and Firefox

(cherry picked from commit dbe4fa6698f5a7086ec322e3e9fd74f71b9ca8f6)