[docs] Update info about eas.json generation and link its config info in the eas.json reference page (#29975)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

Follow-up #29901

As mentioned in the above PR, it was a bit confusing as to when the
**eas.json** file was generated.

# How

<!--
How did you build this feature or fix this bug and why?
-->

This PR:
- Improves verbiage in Configure EAS Build with eas.json when mentioning
how **eas.json** file is generated.
- Update Configure EAS Submit with eas.json to mention how **eas.json**
file is generated. (this page is is not linked with the EAS Build
introduction guide, so based on the feedback received, I think it's okay
to add that context about an **eas.json** file being generated
- Updates **eas.json** reference to link to "Configure EAS Build with
eas.json" and "Configure EAS Submit with eas.json" guides.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

By running docs manually.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).